mpi/init: split MPIR implementations from MPI functions
