+++ src/mpi/init/Makefile.mk
@@ -14,6 +14,9 @@ mpi_sources +=                 \
     src/mpi/init/querythread.c
 
 mpi_core_sources += \
+    src/mpi/init/init_impl.c      \
+    src/mpi/init/mpir_init.c      \
+    src/mpi/init/mpir_finalize.c  \
     src/mpi/init/globals.c        \
     src/mpi/init/initinfo.c       \
     src/mpi/init/local_proc_attrs.c \
