+++ src/mpi/init/abort.c
@@ -8,21 +8,6 @@
 #include <stdlib.h>
 #endif
 
-/*
-=== BEGIN_MPI_T_CVAR_INFO_BLOCK ===
-
-cvars:
-    - name        : MPIR_CVAR_SUPPRESS_ABORT_MESSAGE
-      category    : ERROR_HANDLING
-      type        : boolean
-      default     : false
-      class       : none
-      verbosity   : MPI_T_VERBOSITY_USER_BASIC
-      scope       : MPI_T_SCOPE_ALL_EQ
-      description : Disable printing of abort error message.
-
-=== END_MPI_T_CVAR_INFO_BLOCK ===
-*/
 
 /* -- Begin Profiling Symbol Block for routine MPI_Abort */
 #if defined(HAVE_PRAGMA_WEAK)
@@ -41,7 +26,6 @@ int MPI_Abort(MPI_Comm comm, int errorcode) __attribute__ ((weak, alias("PMPI_Ab
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Abort
 #define MPI_Abort PMPI_Abort
-
 #endif
 
 /*@
@@ -69,7 +53,6 @@ int MPI_Abort(MPI_Comm comm, int errorcode)
 {
     int mpi_errno = MPI_SUCCESS;
     MPIR_Comm *comm_ptr = NULL;
-    int len = MPI_MAX_OBJECT_NAME;
     MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_ABORT);
 
     MPIR_ERRTEST_INITIALIZED_ORDIE();
@@ -112,27 +95,9 @@ int MPI_Abort(MPI_Comm comm, int errorcode)
 
     /* ... body of routine ...  */
 
-    if (!comm_ptr) {
-        /* Use comm world if the communicator is not valid */
-        comm_ptr = MPIR_Process.comm_self;
-    }
-
-    char abort_str[MPI_MAX_OBJECT_NAME + 100] = "";
-    char comm_name[MPI_MAX_OBJECT_NAME];
-    MPIR_Comm_get_name_impl(comm_ptr, comm_name, &len);
-    if (len == 0) {
-        MPL_snprintf(comm_name, MPI_MAX_OBJECT_NAME, "comm=0x%X", comm);
-    }
-    if (!MPIR_CVAR_SUPPRESS_ABORT_MESSAGE)
-        /* FIXME: This is not internationalized */
-        MPL_snprintf(abort_str, sizeof(abort_str),
-                     "application called MPI_Abort(%s, %d) - process %d", comm_name, errorcode,
-                     comm_ptr->rank);
-    mpi_errno = MPID_Abort(comm_ptr, mpi_errno, errorcode, abort_str);
-    /* --BEGIN ERROR HANDLING-- */
+    mpi_errno = MPIR_Abort_impl(comm_ptr, errorcode);
     if (mpi_errno != MPI_SUCCESS)
         goto fn_fail;
-    /* --END ERROR HANDLING-- */
     /* ... end of body of routine ... */
 
   fn_exit:
